@@ -1,6 +1,13 @@
 
 import click
-import pkg_resources
+import attrs
+from typing import Optional
+from ipaddress import IPv4Address, IPv6Address
+from importlib import resources
+from rich.style import Style
+from rich.table import Table
+from rich.text import Text
+from rich.console import Console
 
 from twisted.internet.task import react
 from twisted.internet.defer import ensureDeferred
@@ -15,20 +22,29 @@ from ._proto import (
     frontend_accept_or_invite,
     WELL_KNOWN_MAILBOXES,
 )
-from ._tui import frontend_tui
 from .messages import (
     LocalListener,
     RemoteListener,
 )
-from .policy import (
-    LocalhostTcpPortsListenPolicy,
-    LocalhostTcpPortsConnectPolicy,
-    ArbitraryAddressTcpConnectPolicy,
-    ArbitraryInterfaceTcpPortsListenPolicy,
-    is_localhost,
+from ._tui import frontend_tui
+from .visual import littlebitspace_big_logo
+
+
+WOULD_DO_NOTHING_ERROR = """
+You have requested no listeners and allowed neither listening nor connecting.
+This would not do anything useful.
+
+You should use at least one of: --service or --client (or --remote or --local)
+For more information: fowl --help
+"""
+
+
+
+@click.option(
+    "--version",
+    flag_value=True,
+    help="Output currently running versions as JSON.",
 )
-
-
 @click.option(
     "--ip-privacy/--clearnet",
     default=False,
@@ -48,21 +64,24 @@ from .policy import (
 )
 @click.command()
 @click.pass_context
-def fowld(ctx, ip_privacy, mailbox, debug):
+def fowld(ctx, ip_privacy, mailbox, debug, version):
     """
     Forward Over Wormhole, Locally, Daemon
 
     Low-level daemon to set up and forward streams over Dilated magic
     wormhole connections
     """
+    if version:
+        import json
+        versions = _versions()
+        # versions["kind"] = "version"
+        print(json.dumps(versions))
+        return
+
     ctx.obj = _Config(
         relay_url=WELL_KNOWN_MAILBOXES.get(mailbox, mailbox),
         use_tor=bool(ip_privacy),
         debug_file=debug,
-        # these will be empty; client must activate ports by sending
-        # messages to allow listening (or connecting)
-        listen_policy = LocalhostTcpPortsListenPolicy([]),
-        connect_policy = LocalhostTcpPortsConnectPolicy([]),
     )
     def run(reactor):
         return ensureDeferred(
@@ -74,6 +93,17 @@ def fowld(ctx, ip_privacy, mailbox, debug):
     return react(run)
 
 
+@click.option(
+    "--version",
+    flag_value=True,
+    help="Display the currently running version",
+)
+@click.option(
+    "--debug-messages",
+    default=None,
+    type=click.File(mode="w", encoding="utf8"),
+    help="Save all input/output messages to a file",
+)
 @click.option(
     "--ip-privacy/--clearnet",
     default=False,
@@ -92,43 +122,50 @@ def fowld(ctx, ip_privacy, mailbox, debug):
     type=click.File("w", encoding="utf8"),
 )
 @click.option(
+    "--debug-status",
+    default=None,
+    help="Output the state of FowlStatus (for use with --replay)",
+    type=click.File("w", encoding="utf8", ),
+)
+@click.option(
+    "--replay",
+    help="Replay a series of events recorded with --debug-status",
+    type=click.File("r", encoding="utf8"),
+)
+#XXX big change, no longer supports https://github.com/meejah/fowl/issues/37
+# -> maybe keep these options, and do the 'name'-based ones as a new thing (--daemon-here or --daemon-there)
+# -> maybe do name-based still, but also allow interface (for forwarding...so only on the side that does connect)
+#
+# NOTE also original usage of "--local" was that the 'fake listener' was/is here!
+# (maybe alias --daemon-there / --daemon-here to --local / --remote? --daemon-there == --local (old-style usage))
+@click.option(
     "--local", "-L",
+    "--client",
     multiple=True,
     help=(
-        "Listen locally, connect remotely (accepted multiple times)."
-        "Unless otherwise specified, (local) bind and (remote) connect addresses are localhost."
-        'For example "127.0.0.1:1234:127.0.0.1:22" is the same as "1234:22" effectively.'
+        "We will listen locally, forwarding local connections to the other peer."
+        "That is, the other peer is running the daemon-style software."
+        "The other peer must enable the same service-name."
+        "If a \"remote-connect=\" port is specified, the invocation on the other peer must agree."
+        "Therefore, it is best to ONLY choose ports on your side, unless the protocol requires otherwise."
+        "If you can avoid choosing at all, a random port is assigned -- this is the most likely to succeed."
+        "(Run a corresponding --remote with the same service-name on the other peer)"
     ),
-    metavar="[bind-address:]listen-port[:remote-address][:connect-port]",
+    metavar="service-name:[local-listen-port]:[bind=127.0.0.1]:[remote-connect=port]",
 )
 @click.option(
     "--remote", "-R",
+    "--service",
     multiple=True,
     help=(
-        "Listen remotely, connect locally (accepted multiple times)"
-        "Unless otherwise specified, the (remote) bind and (local) connect addresses are localhost."
-        'For example "127.0.0.1:1234:127.0.0.1:22" is the same as "1234:22" effectively.'
+        "Listen on the other peer, so the server-style software runs here."
+        "The other peer must enable the same service-name."
+        "Ports must agree."
+        "Therefore, it is best to ONLY choose ports on your side, unless the protocol requires otherwise."
+        "If you can avoid choosing at all, a random port is assigned -- this is the most likely to succeed."
+       "(Run a corresponding --local with the same service-name on the other peer)"
         ),
-    metavar="[remote-bind-address:]listen-port[:local-connect-address][:local-connect-port]",
-)
-@click.option(
-    "--allow-listen",
-    multiple=True,
-    help=(
-        "Accept a connection to this local port. Accepted multiple times."
-        "Note that local listeners added via --local are already allowed and do not need this option."
-        'If no interface is specified, "localhost" is assumed.'
-    ),
-    metavar="[interface:]listen-port",
-)
-@click.option(
-    "--allow-connect",
-    multiple=True,
-    help=(
-        "Accept a connection to this local port. Accepted multiple times"
-        'If no address is specified, "localhost" is assumed.'
-    ),
-    metavar="[address:]connect-port",
+    metavar="service-name:[local-connect-port][:listen=port][:address=127.0.0.1]",
 )
 @click.option(
     "--code-length",
@@ -145,9 +182,15 @@ def fowld(ctx, ip_privacy, mailbox, debug):
     help="Run in interactive mode, a human-friendly fowld",
     is_flag=True,
 )
+@click.option(
+    "--no-logo", "--logo",
+    is_flag=True,
+    help="Display the ANSI logo from littlebitspace (or not)",
+    default=False,
+)
 @click.argument("code", required=False)
 @click.command()
-def fowl(ip_privacy, mailbox, debug, allow_listen, allow_connect, local, remote, code_length, code, readme, interactive):
+def fowl(ip_privacy, mailbox, debug, local, remote, code_length, code, readme, interactive, debug_messages, debug_status, replay, version, no_logo):
     """
     Forward Over Wormhole, Locally
 
@@ -160,83 +203,74 @@ def fowl(ip_privacy, mailbox, debug, allow_listen, allow_connect, local, remote,
     This will create a new session (allocating a fresh code) by
     default. To join an existing session (e.g. you've been given a
     code) add the code as an (optional) argument on the command-line.
+
+    This only forwards named services; if *this* peer uses '--local
+    foo' then the other peer must use '--remote foo' and
+    vice-versa. Requesting a port must have corresponding 'permission'
+    on the other side. For example:
+
+        fowl --local chat:4444:1234
+
+    ...must have a corresponding invocation with the *exact* same ports:
+
+        fowl --remote chat:1234:4444
+
+    Then, the first peer can run its own listening software (e.g. "nc
+    -l 4444") and the second peer can run connect-style software
+    (e.g. "telnet localhost 1234")
+
+    We encourge specifying as little information as possible, with the
+    minimum viable setup being just the service names. This will
+    result in random ports (revealed only to the respective peer in
+    their UI):
+
+        fowl --local chat
+        fowl --remote chat
+
+    This form of invocation has the best chance of succeeding, as
+    unused ports are chosen. The first peer still runs listening style
+    softare (e.g. 'nc'), but must retrieve the exact port from the
+    UI. Similarly, the second peer still runs connect style software
+    (e.g. 'telnet localhost') but also finds the exact port from their
+    UI. In this way, the peers don't know which port the other side is
+    actually listening on.
+
+    Note that this can fail for things like Web servers which include
+    the port as part of the URI and the 'same-origin' check.
     """
+    if version:
+        versions = _versions()
+        print(f"    fowl: {versions['fowl']}")
+        print(f"wormhole: {versions['wormhole']}")
+        return
+
+    if replay:
+        _replay_visuals(None, replay)
+        return
+
     if readme:
         display_readme()
         return
 
-    local_commands = [
-        _specifier_to_tuples(cmd)
+    local_services = [
+        LocalSpecifier.parse(cmd)
         for cmd in local
     ]
-    remote_commands = [
-        _specifier_to_tuples(cmd)
+    remote_services = [
+        RemoteSpecifier.parse(cmd)
         for cmd in remote
     ]
 
-    def to_local(local_interface, local_port, remote_address, remote_port):
-        return LocalListener(
-            f"tcp:{local_port}:interface={local_interface}",
-            f"tcp:{remote_address}:{remote_port}",
-        )
+    commands = [
+        spec.to_local()
+        for spec in local_services
+    ] + [
+        spec.to_remote()
+        for spec in remote_services
+    ]
 
-    def to_remote(local_interface, local_port, remote_address, remote_port):
-        return RemoteListener(
-            f"tcp:{local_port}:interface={local_interface}",
-            f"tcp:{remote_address}:{remote_port}",
-        )
-
-    def to_listen_policy(local_interface, local_port, remote_address, remote_port):
-        return local_port
-
-    def to_connect_policy(local_interface, local_port, remote_address, remote_port):
-        return remote_port
-
-    def to_iface_port(allowed):
-        if ':' in allowed:
-            iface, port = allowed.split(':', 1)
-            return iface, _to_port(port)
-        return "localhost", _to_port(allowed)
-
-    def to_local_port(allowed):
-        if ':' in allowed:
-            iface, port = allowed.split(':', 1)
-            if iface != "localhost":
-                raise ValueError(f"Non-local interface: {iface}")
-            return _to_port(port)
-        return _to_port(allowed)
-
-    def is_local(local_interface, local_port, remote_address, remote_port):
-        return is_localhost(local_interface)
-
-    def is_local_connect(local_interface, local_port, remote_address, remote_port):
-        return is_localhost(remote_address)
-
-    if any(not is_local(*cmd) for cmd in local_commands) or \
-       any(not is_localhost(to_iface_port(allowed)[0]) for allowed in allow_listen):
-        listen_policy = ArbitraryInterfaceTcpPortsListenPolicy(
-            [(iface, port) for iface, port, _, _ in local_commands] + \
-            [to_iface_port(allowed) for allowed in allow_listen]
-        )
-    else:
-        listen_policy = LocalhostTcpPortsListenPolicy(
-            [to_listen_policy(*cmd) for cmd in local_commands] +
-            [to_local_port(port) for port in allow_listen]
-        )
-
-    if any(not is_local_connect(*cmd) for cmd in remote_commands) or \
-       any(not is_localhost(to_iface_port(allowed)[0]) for allowed in allow_connect):
-        # yes, this says "to_iface_port()" below but they both look
-        # the same currently: "192.168.1.2:4321" for example
-        connect_policy = ArbitraryAddressTcpConnectPolicy(
-            [(addr, port) for _, _, addr, port in remote_commands] + \
-            [to_iface_port(allowed) for allowed in allow_connect]
-        )
-    else:
-        connect_policy = LocalhostTcpPortsConnectPolicy(
-            [to_connect_policy(*cmd) for cmd in remote_commands] +
-            [to_local_port(port) for port in allow_connect]
-        )
+    if not commands and not interactive:
+        raise click.UsageError(WOULD_DO_NOTHING_ERROR)
 
     cfg = _Config(
         relay_url=WELL_KNOWN_MAILBOXES.get(mailbox, mailbox),
@@ -244,15 +278,10 @@ def fowl(ip_privacy, mailbox, debug, allow_listen, allow_connect, local, remote,
         debug_file=debug,
         code=code,
         code_length=code_length,
-        commands=[
-            to_local(*t)
-            for t in local_commands
-        ] + [
-            to_remote(*t)
-            for t in remote_commands
-        ],
-        listen_policy=listen_policy,
-        connect_policy=connect_policy,
+        commands=commands,
+        output_debug_messages=debug_messages,
+        output_status=debug_status,
+        no_logo=no_logo,
     )
 
     if interactive:
@@ -264,6 +293,8 @@ def fowl(ip_privacy, mailbox, debug, allow_listen, allow_connect, local, remote,
 
 
 def _to_port(arg):
+    if arg is None:
+        return None  # facility defaults, for Specifier parsers
     arg = int(arg)
     if arg < 1 or arg >= 65536:
         raise click.UsageError(
@@ -272,79 +303,200 @@ def _to_port(arg):
     return arg
 
 
-# XXX FIXME use an @frozen attr, not tuple for returns
-def _specifier_to_tuples(cmd):
-    """
-    Parse a local or remote listen/connect specifiers.
+@attrs.frozen
+class RemoteSpecifier:
+    # corresponds to roost()
+    name: str
+    connect_port: Optional[int] = None
+    local_listen_port: Optional[int] = None
+    connect_address: Optional[IPv4Address|IPv6Address] = None
 
-    This always returns a 4-tuple of:
-      - listen interface
-      - listen port
-      - connect address
-      - connect port
-
-    TODO: tests, and IPv6
-    """
-    if '[' in cmd or ']' in cmd:
-        raise RuntimeError("Have not considered IPv6 parsing yet")
-
-    colons = cmd.count(':')
-    if colons > 3:
-        raise ValueError(
-            f"Too many colons: {colons} > 3"
+    def to_remote(self):
+        return RemoteListener(
+            self.name,
+            self.local_listen_port,
+            self.connect_port,
+            self.connect_address,
         )
-    if colons == 3:
-        # everything is specified
-        listen_interface, listen_port, connect_address, connect_port = cmd.split(':')
-        listen_port = _to_port(listen_port)
-        connect_port = _to_port(connect_port)
-    elif colons == 2:
-        # one of the interface / address is specified, but we're not
-        # sure which yet
-        a, b, c = cmd.split(':')
-        try:
-            # maybe the first thing is a port
-            listen_port = _to_port(a)
-            listen_interface = "localhost"
-            connect_address = b
-            connect_port = _to_port(c)
-        except ValueError:
-            # no, the first thing is a string, so the connect address
-            # must be missing
-            listen_interface = a
-            listen_port = _to_port(b)
-            connect_address = "localhost"
-            connect_port = _to_port(c)
-    elif colons == 1:
-        # we only have one split. this could be "interface:port" or "port:port"
-        a, b = cmd.split(':')
-        try:
-            listen_port = _to_port(a)
-            listen_interface = "localhost"
-            try:
-                # the second thing could be a connect address or a
-                # port
-                connect_port = _to_port(b)
-                connect_address = "localhost"
-            except ValueError:
-                connect_address = b
-                connect_port = listen_port
-        except ValueError:
-            # okay, first thing isn't a port so it's the listen interface
-            listen_interface = a
-            listen_port = connect_port = _to_port(b)
-            connect_address = "localhost"
-    else:
-        # no colons, it's a port and we're "symmetric"
-        listen_port = connect_port = _to_port(cmd)
-        listen_interface = "localhost"
-        connect_address = "localhost"
 
-    # XXX ipv6?
-    return (
-        listen_interface, listen_port,
-        connect_address, connect_port,
-    )
+    @staticmethod
+    def parse(cmd: str):
+        if '[' in cmd or ']' in cmd:
+            raise RuntimeError("Have not considered IPv6 parsing yet")
+
+        colons = cmd.count(':')
+        if colons > 3:
+            raise ValueError(
+                f"Too many colons: {colons} > 3"
+            )
+
+        if colons == 0:
+            return RemoteSpecifier(cmd)
+
+        if colons == 1:
+            name, port0 = cmd.split(':')
+            return RemoteSpecifier(name, _to_port(port0))
+
+        specs = cmd.split(':')
+        name = specs.pop(0)
+        port0 = _to_port(specs.pop(0))
+
+        named = {
+            "listen": None,
+            "address": None,
+        }
+        for spec in specs:
+            n, v = spec.split('=')
+            named[n] = v
+            if n not in ["listen", "address"]:
+                raise click.UsageError(
+                    "--remote specifier accepts listen= or address= only"
+                )
+        return RemoteSpecifier(
+            name, port0,
+            local_listen_port=_to_port(named["listen"]),
+            connect_address=named["address"],  # should be IPv{4,6}Address
+        )
+
+
+@attrs.frozen
+class LocalSpecifier:
+    # corresponds to fledge()
+    name: str
+    local_listen_port: Optional[int] = None
+    remote_connect_port: Optional[int] = None
+    bind_interface: Optional[IPv4Address | IPv6Address] = None
+
+    def to_local(self):
+        return LocalListener(
+            self.name,
+            self.local_listen_port,
+            self.remote_connect_port,
+            self.bind_interface,
+        )
+
+    @staticmethod
+    def parse(cmd: str):
+        if '[' in cmd or ']' in cmd:
+            raise RuntimeError("Have not considered IPv6 parsing yet")
+
+        colons = cmd.count(':')
+        if colons > 3:
+            raise ValueError(
+                f"Too many colons: {colons} > 3"
+            )
+
+        if colons == 0:
+            return LocalSpecifier(cmd)
+
+        if colons == 1:
+            name, port0 = cmd.split(':')
+            return LocalSpecifier(name, _to_port(port0))
+
+        specs = cmd.split(':')
+        name = specs.pop(0)
+        port0 = _to_port(specs.pop(0))
+
+        named = {
+            "remote-connect": None,
+            "bind": None,
+        }
+        for spec in specs:
+            n, v = spec.split('=')
+            named[n] = v
+            if n not in ["remote-connect", "bind"]:
+                raise click.UsageError(
+                    "--local specifier accepts remote-connect= or bind= only"
+                )
+        return LocalSpecifier(
+            name, port0,
+            remote_connect_port=_to_port(named["remote-connect"]),
+            bind_interface=named["bind"],  # should be IPv{4,6}Address
+        )
+
+
+Specifier = LocalSpecifier | RemoteSpecifier
+
+
+def _replay_visuals(cfg, messages):
+    """
+    Replay a series of events output from --debug-messages
+
+    Such a file consists of one JSON message per line; the timestamp=
+    values are honoured, so the playback is real time.
+    """
+    from rich.live import Live
+    from .visual import render_status
+    from .status import _StatusTracker, FowlStatus, Listener, Subchannel
+    import time
+    import json
+
+
+    with messages as f:
+        messages = [
+            json.loads(line)
+            for line in f.readlines()
+        ]
+    where_are_we = messages[0]["timestamp"]
+
+    def current_time():
+        print(f"current {where_are_we}")
+        return where_are_we
+    status_tracker = _StatusTracker(time_provider=current_time)
+
+    def render():
+        return render_status(status_tracker.current_status, where_are_we)
+    from rich.console import Console
+    console = Console(force_terminal=True)
+    live = Live(get_renderable=render, console=console)
+
+    #if 1:
+    with live:
+        while messages:
+            data = messages.pop(0)
+            timestamp = data.pop("timestamp")
+
+            msg = FowlStatus(**data)
+            msg = attrs.evolve(
+                msg,
+                listeners={
+                    kw["service_name"]: Listener(**kw)
+                    for kw in msg.listeners.values()
+                },
+                subchannels={
+                    kw["channel_id"]: Subchannel(**kw)
+                    for kw in msg.subchannels.values()
+                },
+            )
+            print(msg)
+            # time is hard
+            # intuitively, we want to trigger a redraw 4 times a second
+            # ...but waiting 0.25s with time.sleep() isn't right,
+            # because we take a non-zero amount of time to do the rest
+            # ... and are we doing "wall-clock time" or some kind of
+            # virtual shit here, who can say.
+            delay = timestamp - where_are_we
+            target = time.time() + delay
+            while time.time() < target:
+                status_tracker._current_status = msg
+                amount = min(0.23, max(0.0, target - time.time()))
+                where_are_we += amount
+                time.sleep(amount)
+
+            # regardless of any overruns, clam to target
+            where_are_we = timestamp
+
+
+def _versions():
+    """
+    Return information about interesting versions
+    """
+    from fowl import __version__ as fowl_version
+    from wormhole import __version__ as wormhole_version
+    return {
+        "fowl": fowl_version,
+        "wormhole": wormhole_version,
+    }
 
 
 def tui(cfg):
@@ -364,12 +516,21 @@ def display_readme():
     """
     Display the project README
     """
-    readme = pkg_resources.resource_string('fowl', '../../README.rst')
-    # uhm, docutils documentation is confusing as all hell and no good
-    # examples of "convert this rST string to anything else" .. :/ but
-    # we should "render" it to text
-    click.echo_via_pager(readme.decode('utf8'))
+    layout = Table.grid("text", "digram")
+    logo = Text.from_ansi(
+        littlebitspace_big_logo,
+        no_wrap=True,
+        style=Style(bgcolor="#002b36"),
+    )
+    layout.add_row(
+        resources.files("fowl").joinpath("README.rst").read_text(encoding="utf8"),
+        logo,
+    )
 
+    c = Console()
+    with c.capture() as capture:
+        c.print(layout)
+    click.echo_via_pager(capture.get())
 
 if __name__ == "__main__":
     try:
