@@ -1,1 +1,1 @@
-__version__ = "24.12.0"
+__version__ = "25.7.0"
